#67 - Cleanup StateMachine.cs and rename workflow titles
