+++ .github/ISSUE_TEMPLATE/1-feature.yaml
@@ -1,6 +1,6 @@
 name: Feature Request
 description: Request a new feature or enhancement.
-title: "StateCraft: "
+title: "Feature: "
 labels: []
 projects: ["ZCrewSoftware/ZCrew.StateCraft"]
 type: feature
