+++ .github/ISSUE_TEMPLATE/2-bugfix.yaml
@@ -1,6 +1,6 @@
 name: Bug Report
 description: File a bug report.
-title: "StateCraft: "
+title: "Bug: "
 labels: []
 projects: ["ZCrewSoftware/ZCrew.StateCraft"]
 type: bugfix
