+++ .github/ISSUE_TEMPLATE/4-spike.yaml
@@ -1,6 +1,6 @@
 name: Spike
 description: Exploratory work to create a proof of concept
-title: "StateCraft: "
+title: "Spike: "
 labels: []
 projects: ["ZCrewSoftware/ZCrew.StateCraft"]
 type: spike
