+++ src/ZCrew.StateCraft/StateMachines/StateMachine.cs
@@ -2,6 +2,7 @@ using System.Diagnostics;
 using System.Diagnostics.CodeAnalysis;
 using Nito.AsyncEx;
 using ZCrew.Extensions.Tasks;
+using ZCrew.StateCraft.Extensions;
 using ZCrew.StateCraft.Parameters;
 using ZCrew.StateCraft.Parameters.Contracts;
 using ZCrew.StateCraft.StateMachines.Contracts;
@@ -13,7 +14,7 @@ using ZCrew.StateCraft.Triggers.Contracts;
 namespace ZCrew.StateCraft.StateMachines;
 
 /// <inheritdoc />
-internal sealed class StateMachine<TState, TTransition> : IStateMachine<TState, TTransition>
+internal sealed partial class StateMachine<TState, TTransition> : IStateMachine<TState, TTransition>
     where TState : notnull
     where TTransition : notnull
 {
@@ -75,7 +76,116 @@ internal sealed class StateMachine<TState, TTransition> : IStateMachine<TState, 
     public IStateMachineParameters Parameters { get; } = new StateMachineParameters();
 
     /// <inheritdoc />
-    public ITransition<TState, TTransition>? CurrentTransition { get; set; }
+    public async Task Activate(CancellationToken token = default)
+    {
+        using var activationLock = await this.stateMachineLock.LockAsync(token);
+        if (this.internalState.IsActivated)
+        {
+            throw new InvalidOperationException("The state machine has already been activated.");
+        }
+
+        try
+        {
+            await this.stateMachineActivator.Activate(this, token);
+            Debug.Assert(NextState != null, $"Expected {nameof(NextState)} to be set.");
+
+            await NextState.Activate(Parameters, token);
+            this.internalState = InternalState.Idle;
+            await ActivateTriggers(token);
+            await EnterState(activationLock, token);
+        }
+        catch when (CurrentState == null)
+        {
+            // If there was an exception during activation the state machine is forced to reset fully
+            Parameters.Clear();
+            NextState = null;
+            CurrentState = null;
+            PreviousState = null;
+
+            this.internalState = InternalState.Inactive;
+            await DeactivateTriggers(CancellationToken.None);
+            throw;
+        }
+        catch
+        {
+            await DeactivateTriggers(CancellationToken.None);
+            throw;
+        }
+    }
+
+    /// <inheritdoc />
+    public async Task Deactivate(CancellationToken token = default)
+    {
+        using var _ = await this.stateMachineLock.LockAsync(token);
+        if (this.internalState.IsInactive)
+        {
+            throw new InvalidOperationException("The state machine has not been activated.");
+        }
+
+        if (this.internalState.IsEntering)
+        {
+            await DeactivateTriggers(token);
+            Parameters.Clear();
+            PreviousState = null;
+            NextState = null;
+            this.internalState = InternalState.Inactive;
+            return;
+        }
+
+        try
+        {
+            BeginTransition();
+            await ExitState(token);
+            await DeactivateTriggers(token);
+
+            await PreviousState.Deactivate(Parameters, token);
+
+            Parameters.Clear();
+            PreviousState = null;
+            this.internalState = InternalState.Inactive;
+        }
+        catch when (CurrentState == null)
+        {
+            Rollback();
+            this.internalState = InternalState.Recovery;
+            throw;
+        }
+    }
+
+    /// <inheritdoc />
+    public async Task StateChange(
+        TState previousState,
+        TTransition transition,
+        TState nextState,
+        CancellationToken token
+    )
+    {
+        foreach (var onStateChange in this.onStateChanges)
+        {
+            await ExceptionBehavior.CallOnStateChange(
+                t => onStateChange.InvokeAsync(previousState, transition, nextState, t),
+                token
+            );
+        }
+    }
+
+    /// <inheritdoc />
+    public void AddState(IState<TState, TTransition> state)
+    {
+        StateTable.Add(state);
+    }
+
+    /// <inheritdoc />
+    public void Dispose()
+    {
+        var actionCts = this.actionCancellationTokenSource;
+
+        this.actionTask = null;
+        this.actionCancellationTokenSource = null;
+
+        actionCts?.Cancel();
+        actionCts?.Dispose();
+    }
 
     private async Task ExitState(CancellationToken token)
     {
@@ -109,17 +219,14 @@ internal sealed class StateMachine<TState, TTransition> : IStateMachine<TState, 
         this.internalState = InternalState.Exited;
     }
 
-    private async Task Transition(CancellationToken token)
+    private async Task ExecuteTransition(ITransition<TState, TTransition> currentTransition, CancellationToken token)
     {
         Debug.Assert(PreviousState != null, $"Expected {nameof(PreviousState)} to be set.");
-        Debug.Assert(CurrentTransition != null, $"Expected {nameof(CurrentTransition)} to be set.");
         Debug.Assert(NextState != null, $"Expected {nameof(NextState)} to be set.");
 
         this.internalState = InternalState.Transitioning;
-        await CurrentTransition.Transition(Parameters, token);
+        await currentTransition.Transition(Parameters, token);
         this.internalState = InternalState.Transitioned;
-
-        CurrentTransition = null;
     }
 
     /// <param name="methodLock">The state machine lock to release.</param>
@@ -190,8 +297,6 @@ internal sealed class StateMachine<TState, TTransition> : IStateMachine<TState, 
         PreviousState = CurrentState;
         CurrentState = null;
         NextState = null;
-
-        CurrentTransition = null;
     }
 
     private void Rollback()
@@ -200,8 +305,6 @@ internal sealed class StateMachine<TState, TTransition> : IStateMachine<TState, 
         NextState = null;
         CurrentState = PreviousState;
         PreviousState = null;
-
-        CurrentTransition = null;
     }
 
     /// <summary>
@@ -220,7 +323,6 @@ internal sealed class StateMachine<TState, TTransition> : IStateMachine<TState, 
         PreviousState = null;
         CurrentState = NextState;
         NextState = null;
-        CurrentTransition = null;
         this.internalState = InternalState.Active;
 
         // Use a canceled token here so that any async work is canceled immediately; but synchronous work (or work that
@@ -246,805 +348,9 @@ internal sealed class StateMachine<TState, TTransition> : IStateMachine<TState, 
         }
     }
 
-    /// <inheritdoc />
-    public async Task Activate(CancellationToken token = default)
-    {
-        using var activationLock = await this.stateMachineLock.LockAsync(token);
-        if (this.internalState is not InternalState.Inactive)
-        {
-            throw new InvalidOperationException("The state machine has already been activated.");
-        }
-
-        try
-        {
-            await this.stateMachineActivator.Activate(this, token);
-            Debug.Assert(NextState != null, $"Expected {nameof(NextState)} to be set.");
-
-            await NextState.Activate(Parameters, token);
-            this.internalState = InternalState.Idle;
-            await ActivateTriggers(token);
-            await EnterState(activationLock, token);
-        }
-        catch when (CurrentState == null)
-        {
-            // If there was an exception during activation the state machine is forced to reset fully
-            Parameters.Clear();
-            NextState = null;
-            CurrentState = null;
-            PreviousState = null;
-
-            CurrentTransition = null;
-            this.internalState = InternalState.Inactive;
-            await DeactivateTriggers(CancellationToken.None);
-            throw;
-        }
-        catch
-        {
-            await DeactivateTriggers(CancellationToken.None);
-            throw;
-        }
-    }
-
-    /// <inheritdoc />
-    public async Task Deactivate(CancellationToken token = default)
-    {
-        using var _ = await this.stateMachineLock.LockAsync(token);
-        if (this.internalState is InternalState.Inactive)
-        {
-            throw new InvalidOperationException("The state machine has not been activated.");
-        }
-
-        if (this.internalState is InternalState.Entering)
-        {
-            await DeactivateTriggers(token);
-            Parameters.Clear();
-            PreviousState = null;
-            NextState = null;
-            CurrentTransition = null;
-            this.internalState = InternalState.Inactive;
-            return;
-        }
-
-        try
-        {
-            BeginTransition();
-            await ExitState(token);
-            await DeactivateTriggers(token);
-
-            await PreviousState.Deactivate(Parameters, token);
-
-            Parameters.Clear();
-            PreviousState = null;
-            this.internalState = InternalState.Inactive;
-        }
-        catch when (CurrentState == null)
-        {
-            Rollback();
-            this.internalState = InternalState.Recovery;
-            throw;
-        }
-    }
-
-    /// <inheritdoc />
-    public async Task Transition(TTransition transition, CancellationToken token = default)
-    {
-        using var transitionLock = await this.stateMachineLock.LockAsync(token);
-        if (
-            this.internalState is not InternalState.Active and not InternalState.Recovery and not InternalState.Entering
-        )
-        {
-            throw new InvalidOperationException("The state machine has not been activated.");
-        }
-
-        if (this.internalState is InternalState.Entering)
-        {
-            await RetryEntry(token);
-        }
-
-        try
-        {
-            BeginTransition();
-            Parameters.SetEmptyNextParameters();
-            CurrentTransition = await PreviousState.GetTransition(transition, Parameters, token);
-            NextState = CurrentTransition.Next.State;
-            await ExitState(token);
-            await Transition(token);
-        }
-        catch when (CurrentState == null)
-        {
-            Rollback();
-            this.internalState = InternalState.Recovery;
-            throw;
-        }
-
-        await EnterState(transitionLock, token);
-    }
-
-    /// <inheritdoc />
-    public async Task Transition<T>(TTransition transition, T parameter, CancellationToken token = default)
-    {
-        using var transitionLock = await this.stateMachineLock.LockAsync(token);
-        if (
-            this.internalState is not InternalState.Active and not InternalState.Recovery and not InternalState.Entering
-        )
-        {
-            throw new InvalidOperationException("The state machine has not been activated.");
-        }
-
-        if (this.internalState is InternalState.Entering)
-        {
-            await RetryEntry(token);
-        }
-
-        try
-        {
-            BeginTransition();
-            Parameters.SetNextParameter(parameter);
-            CurrentTransition = await PreviousState.GetTransition(transition, Parameters, token);
-            NextState = CurrentTransition.Next.State;
-            await ExitState(token);
-            await Transition(token);
-        }
-        catch when (CurrentState == null)
-        {
-            Rollback();
-            this.internalState = InternalState.Recovery;
-            throw;
-        }
-
-        await EnterState(transitionLock, token);
-    }
-
-    /// <inheritdoc />
-    public async Task Transition<T1, T2>(
-        TTransition transition,
-        T1 parameter1,
-        T2 parameter2,
-        CancellationToken token = default
-    )
-    {
-        using var transitionLock = await this.stateMachineLock.LockAsync(token);
-        if (
-            this.internalState is not InternalState.Active and not InternalState.Recovery and not InternalState.Entering
-        )
-        {
-            throw new InvalidOperationException("The state machine has not been activated.");
-        }
-
-        if (this.internalState is InternalState.Entering)
-        {
-            await RetryEntry(token);
-        }
-
-        try
-        {
-            BeginTransition();
-            Parameters.SetNextParameters(parameter1, parameter2);
-            CurrentTransition = await PreviousState.GetTransition(transition, Parameters, token);
-            NextState = CurrentTransition.Next.State;
-            await ExitState(token);
-            await Transition(token);
-        }
-        catch when (CurrentState == null)
-        {
-            Rollback();
-            this.internalState = InternalState.Recovery;
-            throw;
-        }
-
-        await EnterState(transitionLock, token);
-    }
-
-    /// <inheritdoc />
-    public async Task Transition<T1, T2, T3>(
-        TTransition transition,
-        T1 parameter1,
-        T2 parameter2,
-        T3 parameter3,
-        CancellationToken token = default
-    )
-    {
-        using var transitionLock = await this.stateMachineLock.LockAsync(token);
-        if (
-            this.internalState is not InternalState.Active and not InternalState.Recovery and not InternalState.Entering
-        )
-        {
-            throw new InvalidOperationException("The state machine has not been activated.");
-        }
-
-        if (this.internalState is InternalState.Entering)
-        {
-            await RetryEntry(token);
-        }
-
-        try
-        {
-            BeginTransition();
-            Parameters.SetNextParameters(parameter1, parameter2, parameter3);
-            CurrentTransition = await PreviousState.GetTransition(transition, Parameters, token);
-            NextState = CurrentTransition.Next.State;
-            await ExitState(token);
-            await Transition(token);
-        }
-        catch when (CurrentState == null)
-        {
-            Rollback();
-            this.internalState = InternalState.Recovery;
-            throw;
-        }
-
-        await EnterState(transitionLock, token);
-    }
-
-    /// <inheritdoc />
-    public async Task Transition<T1, T2, T3, T4>(
-        TTransition transition,
-        T1 parameter1,
-        T2 parameter2,
-        T3 parameter3,
-        T4 parameter4,
-        CancellationToken token = default
-    )
-    {
-        using var transitionLock = await this.stateMachineLock.LockAsync(token);
-        if (
-            this.internalState is not InternalState.Active and not InternalState.Recovery and not InternalState.Entering
-        )
-        {
-            throw new InvalidOperationException("The state machine has not been activated.");
-        }
-
-        if (this.internalState is InternalState.Entering)
-        {
-            await RetryEntry(token);
-        }
-
-        try
-        {
-            BeginTransition();
-            Parameters.SetNextParameters(parameter1, parameter2, parameter3, parameter4);
-            CurrentTransition = await PreviousState.GetTransition(transition, Parameters, token);
-            NextState = CurrentTransition.Next.State;
-            await ExitState(token);
-            await Transition(token);
-        }
-        catch when (CurrentState == null)
-        {
-            Rollback();
-            this.internalState = InternalState.Recovery;
-            throw;
-        }
-
-        await EnterState(transitionLock, token);
-    }
-
-    /// <inheritdoc />
-    public async Task<bool> CanTransition(TTransition transition, CancellationToken token = default)
-    {
-        using var _ = await this.stateMachineLock.LockAsync(token);
-        if (
-            this.internalState is not InternalState.Active and not InternalState.Recovery and not InternalState.Entering
-        )
-        {
-            return false;
-        }
-
-        if (this.internalState is InternalState.Entering)
-        {
-            await RetryEntry(token);
-        }
-
-        try
-        {
-            BeginTransition();
-            Parameters.SetEmptyNextParameters();
-            var nextTransition = await PreviousState.GetTransitionOrDefault(transition, Parameters, token);
-            return nextTransition != null;
-        }
-        finally
-        {
-            Rollback();
-        }
-    }
-
-    /// <inheritdoc />
-    public async Task<bool> CanTransition<T>(TTransition transition, T parameter, CancellationToken token = default)
-    {
-        using var _ = await this.stateMachineLock.LockAsync(token);
-        if (
-            this.internalState is not InternalState.Active and not InternalState.Recovery and not InternalState.Entering
-        )
-        {
-            return false;
-        }
-
-        if (this.internalState is InternalState.Entering)
-        {
-            await RetryEntry(token);
-        }
-
-        try
-        {
-            BeginTransition();
-            Parameters.SetNextParameter(parameter);
-            var nextTransition = await PreviousState.GetTransitionOrDefault(transition, Parameters, token);
-            return nextTransition != null;
-        }
-        finally
-        {
-            Rollback();
-        }
-    }
-
-    /// <inheritdoc />
-    public async Task<bool> CanTransition<T1, T2>(
-        TTransition transition,
-        T1 parameter1,
-        T2 parameter2,
-        CancellationToken token = default
-    )
-    {
-        using var _ = await this.stateMachineLock.LockAsync(token);
-        if (
-            this.internalState is not InternalState.Active and not InternalState.Recovery and not InternalState.Entering
-        )
-        {
-            return false;
-        }
-
-        if (this.internalState is InternalState.Entering)
-        {
-            await RetryEntry(token);
-        }
-
-        try
-        {
-            BeginTransition();
-            Parameters.SetNextParameters(parameter1, parameter2);
-            var nextTransition = await PreviousState.GetTransitionOrDefault(transition, Parameters, token);
-            return nextTransition != null;
-        }
-        finally
-        {
-            Rollback();
-        }
-    }
-
-    /// <inheritdoc />
-    public async Task<bool> CanTransition<T1, T2, T3>(
-        TTransition transition,
-        T1 parameter1,
-        T2 parameter2,
-        T3 parameter3,
-        CancellationToken token = default
-    )
-    {
-        using var _ = await this.stateMachineLock.LockAsync(token);
-        if (
-            this.internalState is not InternalState.Active and not InternalState.Recovery and not InternalState.Entering
-        )
-        {
-            return false;
-        }
-
-        if (this.internalState is InternalState.Entering)
-        {
-            await RetryEntry(token);
-        }
-
-        try
-        {
-            BeginTransition();
-            Parameters.SetNextParameters(parameter1, parameter2, parameter3);
-            var nextTransition = await PreviousState.GetTransitionOrDefault(transition, Parameters, token);
-            return nextTransition != null;
-        }
-        finally
-        {
-            Rollback();
-        }
-    }
-
-    /// <inheritdoc />
-    public async Task<bool> CanTransition<T1, T2, T3, T4>(
-        TTransition transition,
-        T1 parameter1,
-        T2 parameter2,
-        T3 parameter3,
-        T4 parameter4,
-        CancellationToken token = default
-    )
-    {
-        using var _ = await this.stateMachineLock.LockAsync(token);
-        if (
-            this.internalState is not InternalState.Active and not InternalState.Recovery and not InternalState.Entering
-        )
-        {
-            return false;
-        }
-
-        if (this.internalState is InternalState.Entering)
-        {
-            await RetryEntry(token);
-        }
-
-        try
-        {
-            BeginTransition();
-            Parameters.SetNextParameters(parameter1, parameter2, parameter3, parameter4);
-            var nextTransition = await PreviousState.GetTransitionOrDefault(transition, Parameters, token);
-            return nextTransition != null;
-        }
-        finally
-        {
-            Rollback();
-        }
-    }
-
-    /// <inheritdoc />
-    public async Task<bool> TryTransition(TTransition transition, CancellationToken token = default)
-    {
-        using var transitionLock = await this.stateMachineLock.LockAsync(token);
-        if (
-            this.internalState is not InternalState.Active and not InternalState.Recovery and not InternalState.Entering
-        )
-        {
-            return false;
-        }
-
-        if (this.internalState is InternalState.Entering)
-        {
-            await RetryEntry(token);
-        }
-
-        try
-        {
-            BeginTransition();
-            Parameters.SetEmptyNextParameters();
-            CurrentTransition = await PreviousState.GetTransitionOrDefault(transition, Parameters, token);
-            if (CurrentTransition == null)
-            {
-                Rollback();
-                return false;
-            }
-        }
-        catch
-        {
-            Rollback();
-            throw;
-        }
-
-        try
-        {
-            NextState = CurrentTransition.Next.State;
-            await ExitState(token);
-            await Transition(token);
-        }
-        catch when (CurrentState == null)
-        {
-            Rollback();
-            this.internalState = InternalState.Recovery;
-            throw;
-        }
-
-        await EnterState(transitionLock, token);
-        return true;
-    }
-
-    /// <inheritdoc />
-    public async Task<bool> TryTransition<T>(TTransition transition, T parameter, CancellationToken token = default)
-    {
-        using var transitionLock = await this.stateMachineLock.LockAsync(token);
-        if (
-            this.internalState is not InternalState.Active and not InternalState.Recovery and not InternalState.Entering
-        )
-        {
-            return false;
-        }
-
-        if (this.internalState is InternalState.Entering)
-        {
-            await RetryEntry(token);
-        }
-
-        try
-        {
-            BeginTransition();
-            Parameters.SetNextParameter(parameter);
-            CurrentTransition = await PreviousState.GetTransitionOrDefault(transition, Parameters, token);
-            if (CurrentTransition == null)
-            {
-                Rollback();
-                return false;
-            }
-        }
-        catch
-        {
-            Rollback();
-            throw;
-        }
-
-        try
-        {
-            NextState = CurrentTransition.Next.State;
-            await ExitState(token);
-            await Transition(token);
-        }
-        catch when (CurrentState == null)
-        {
-            Rollback();
-            this.internalState = InternalState.Recovery;
-            throw;
-        }
-
-        await EnterState(transitionLock, token);
-        return true;
-    }
-
-    /// <inheritdoc />
-    public async Task<bool> TryTransition<T1, T2>(
-        TTransition transition,
-        T1 parameter1,
-        T2 parameter2,
-        CancellationToken token = default
-    )
-    {
-        using var transitionLock = await this.stateMachineLock.LockAsync(token);
-        if (
-            this.internalState is not InternalState.Active and not InternalState.Recovery and not InternalState.Entering
-        )
-        {
-            return false;
-        }
-
-        if (this.internalState is InternalState.Entering)
-        {
-            await RetryEntry(token);
-        }
-
-        try
-        {
-            BeginTransition();
-            Parameters.SetNextParameters(parameter1, parameter2);
-            CurrentTransition = await PreviousState.GetTransitionOrDefault(transition, Parameters, token);
-            if (CurrentTransition == null)
-            {
-                Rollback();
-                return false;
-            }
-        }
-        catch
-        {
-            Rollback();
-            throw;
-        }
-
-        try
-        {
-            NextState = CurrentTransition.Next.State;
-            await ExitState(token);
-            await Transition(token);
-        }
-        catch when (CurrentState == null)
-        {
-            Rollback();
-            this.internalState = InternalState.Recovery;
-            throw;
-        }
-
-        await EnterState(transitionLock, token);
-        return true;
-    }
-
-    /// <inheritdoc />
-    public async Task<bool> TryTransition<T1, T2, T3>(
-        TTransition transition,
-        T1 parameter1,
-        T2 parameter2,
-        T3 parameter3,
-        CancellationToken token = default
-    )
-    {
-        using var transitionLock = await this.stateMachineLock.LockAsync(token);
-        if (
-            this.internalState is not InternalState.Active and not InternalState.Recovery and not InternalState.Entering
-        )
-        {
-            return false;
-        }
-
-        if (this.internalState is InternalState.Entering)
-        {
-            await RetryEntry(token);
-        }
-
-        try
-        {
-            BeginTransition();
-            Parameters.SetNextParameters(parameter1, parameter2, parameter3);
-            CurrentTransition = await PreviousState.GetTransitionOrDefault(transition, Parameters, token);
-            if (CurrentTransition == null)
-            {
-                Rollback();
-                return false;
-            }
-        }
-        catch
-        {
-            Rollback();
-            throw;
-        }
-
-        try
-        {
-            NextState = CurrentTransition.Next.State;
-            await ExitState(token);
-            await Transition(token);
-        }
-        catch when (CurrentState == null)
-        {
-            Rollback();
-            this.internalState = InternalState.Recovery;
-            throw;
-        }
-
-        await EnterState(transitionLock, token);
-        return true;
-    }
-
-    /// <inheritdoc />
-    public async Task<bool> TryTransition<T1, T2, T3, T4>(
-        TTransition transition,
-        T1 parameter1,
-        T2 parameter2,
-        T3 parameter3,
-        T4 parameter4,
-        CancellationToken token = default
-    )
-    {
-        using var transitionLock = await this.stateMachineLock.LockAsync(token);
-        if (
-            this.internalState is not InternalState.Active and not InternalState.Recovery and not InternalState.Entering
-        )
-        {
-            return false;
-        }
-
-        if (this.internalState is InternalState.Entering)
-        {
-            await RetryEntry(token);
-        }
-
-        try
-        {
-            BeginTransition();
-            Parameters.SetNextParameters(parameter1, parameter2, parameter3, parameter4);
-            CurrentTransition = await PreviousState.GetTransitionOrDefault(transition, Parameters, token);
-            if (CurrentTransition == null)
-            {
-                Rollback();
-                return false;
-            }
-        }
-        catch
-        {
-            Rollback();
-            throw;
-        }
-
-        try
-        {
-            NextState = CurrentTransition.Next.State;
-            await ExitState(token);
-            await Transition(token);
-        }
-        catch when (CurrentState == null)
-        {
-            Rollback();
-            this.internalState = InternalState.Recovery;
-            throw;
-        }
-
-        await EnterState(transitionLock, token);
-        return true;
-    }
-
-    /// <inheritdoc />
-    public async Task StateChange(
-        TState previousState,
-        TTransition transition,
-        TState nextState,
-        CancellationToken token
-    )
-    {
-        foreach (var onStateChange in this.onStateChanges)
-        {
-            await ExceptionBehavior.CallOnStateChange(
-                t => onStateChange.InvokeAsync(previousState, transition, nextState, t),
-                token
-            );
-        }
-    }
-
-    /// <inheritdoc />
-    public void AddState(IState<TState, TTransition> state)
-    {
-        StateTable.Add(state);
-    }
-
-    /// <inheritdoc />
-    public void Dispose()
-    {
-        var actionCts = this.actionCancellationTokenSource;
-
-        this.actionTask = null;
-        this.actionCancellationTokenSource = null;
-
-        actionCts?.Cancel();
-        actionCts?.Dispose();
-    }
-
     [Conditional("DEBUG")]
     private void SetupTracking()
     {
         Tracker = new DebugTracker<TState, TTransition>();
-    }
-
-    /// <summary>
-    ///     The internal state of the state machine.
-    /// </summary>
-    private enum InternalState
-    {
-        /// <summary>
-        ///     The state machine has not been activated.
-        /// </summary>
-        Inactive,
-
-        /// <summary>
-        ///     The state machine is active and ready for transitions.
-        /// </summary>
-        Active,
-
-        /// <summary>
-        ///     The state machine has completed activation and is ready to enter the next state.
-        /// </summary>
-        Idle,
-
-        /// <summary>
-        ///     The state machine is exiting a state.
-        /// </summary>
-        Exiting,
-
-        /// <summary>
-        ///     The state machine has exited the state.
-        /// </summary>
-        Exited,
-
-        /// <summary>
-        ///     The state machine is in the process of transitioning between states.
-        /// </summary>
-        Transitioning,
-
-        /// <summary>
-        ///     The state machine has transitioned: setting the next state, setting the next parameter, and has invoked
-        ///     the state change handlers. The state machine is ready to enter the next state.
-        /// </summary>
-        Transitioned,
-
-        /// <summary>
-        ///     The state machine is entering a state.
-        /// </summary>
-        Entering,
-
-        /// <summary>
-        ///     The state machine has entered the state and the state action is ready to be ran.
-        /// </summary>
-        Entered,
-
-        /// <summary>
-        ///     The state machine failed to transition. It was rolled-back to it's previous state but did not re-enter
-        ///     or restart the actions for that state.
-        /// </summary>
-        Recovery,
     }
 }
