+++ tests/ZCrew.StateCraft.IntegrationTests/ExceptionHandling/ActivateExceptionTests{T1,T2,T3}.cs
@@ -29,7 +29,6 @@ public partial class ActivateExceptionTests
         Assert.False(stateMachine.Parameters.IsCurrentSet);
         Assert.False(stateMachine.Parameters.IsPreviousSet);
         Assert.False(stateMachine.Parameters.IsNextSet);
-        Assert.Null(stateMachine.CurrentTransition);
     }
 
     [Fact]
@@ -91,7 +90,6 @@ public partial class ActivateExceptionTests
         Assert.False(stateMachine.Parameters.IsCurrentSet);
         Assert.False(stateMachine.Parameters.IsPreviousSet);
         Assert.False(stateMachine.Parameters.IsNextSet);
-        Assert.Null(stateMachine.CurrentTransition);
     }
 
     [Fact]
@@ -149,7 +147,6 @@ public partial class ActivateExceptionTests
         Assert.False(stateMachine.Parameters.IsCurrentSet);
         Assert.False(stateMachine.Parameters.IsPreviousSet);
         Assert.False(stateMachine.Parameters.IsNextSet);
-        Assert.Null(stateMachine.CurrentTransition);
     }
 
     [Fact]
