+++ tests/ZCrew.StateCraft.IntegrationTests/ExceptionHandling/ActivateExceptionTests{T1,T2}.cs
@@ -29,7 +29,6 @@ public partial class ActivateExceptionTests
         Assert.False(stateMachine.Parameters.IsCurrentSet);
         Assert.False(stateMachine.Parameters.IsPreviousSet);
         Assert.False(stateMachine.Parameters.IsNextSet);
-        Assert.Null(stateMachine.CurrentTransition);
     }
 
     [Fact]
@@ -88,7 +87,6 @@ public partial class ActivateExceptionTests
         Assert.False(stateMachine.Parameters.IsCurrentSet);
         Assert.False(stateMachine.Parameters.IsPreviousSet);
         Assert.False(stateMachine.Parameters.IsNextSet);
-        Assert.Null(stateMachine.CurrentTransition);
     }
 
     [Fact]
@@ -146,7 +144,6 @@ public partial class ActivateExceptionTests
         Assert.False(stateMachine.Parameters.IsCurrentSet);
         Assert.False(stateMachine.Parameters.IsPreviousSet);
         Assert.False(stateMachine.Parameters.IsNextSet);
-        Assert.Null(stateMachine.CurrentTransition);
     }
 
     [Fact]
