+++ tests/ZCrew.StateCraft.IntegrationTests/ExceptionHandling/ActivateExceptionTests{T}.cs
@@ -29,7 +29,6 @@ public partial class ActivateExceptionTests
         Assert.False(stateMachine.Parameters.IsCurrentSet);
         Assert.False(stateMachine.Parameters.IsPreviousSet);
         Assert.False(stateMachine.Parameters.IsNextSet);
-        Assert.Null(stateMachine.CurrentTransition);
     }
 
     [Fact]
@@ -89,7 +88,6 @@ public partial class ActivateExceptionTests
         Assert.False(stateMachine.Parameters.IsCurrentSet);
         Assert.False(stateMachine.Parameters.IsPreviousSet);
         Assert.False(stateMachine.Parameters.IsNextSet);
-        Assert.Null(stateMachine.CurrentTransition);
     }
 
     [Fact]
@@ -148,7 +146,6 @@ public partial class ActivateExceptionTests
         Assert.False(stateMachine.Parameters.IsCurrentSet);
         Assert.False(stateMachine.Parameters.IsPreviousSet);
         Assert.False(stateMachine.Parameters.IsNextSet);
-        Assert.Null(stateMachine.CurrentTransition);
     }
 
     [Fact]
