+++ tests/ZCrew.StateCraft.IntegrationTests/ExceptionHandling/TransitionExceptionTests{T1,T2,T3,T4}.cs
@@ -36,7 +36,6 @@ public class TransitionExceptionTests_T1_T2_T3_T4
         Assert.Empty(stateMachine.Parameters.CurrentParameterTypes);
         Assert.False(stateMachine.Parameters.IsPreviousSet);
         Assert.False(stateMachine.Parameters.IsNextSet);
-        Assert.Null(stateMachine.CurrentTransition);
     }
 
     [Fact]
@@ -142,7 +141,6 @@ public class TransitionExceptionTests_T1_T2_T3_T4
         Assert.Empty(stateMachine.Parameters.CurrentParameterTypes);
         Assert.False(stateMachine.Parameters.IsPreviousSet);
         Assert.False(stateMachine.Parameters.IsNextSet);
-        Assert.Null(stateMachine.CurrentTransition);
     }
 
     [Fact]
@@ -254,7 +252,6 @@ public class TransitionExceptionTests_T1_T2_T3_T4
         Assert.Empty(stateMachine.Parameters.CurrentParameterTypes);
         Assert.False(stateMachine.Parameters.IsPreviousSet);
         Assert.False(stateMachine.Parameters.IsNextSet);
-        Assert.Null(stateMachine.CurrentTransition);
     }
 
     [Fact]
@@ -383,7 +380,6 @@ public class TransitionExceptionTests_T1_T2_T3_T4
         Assert.Equal((42, "hello", 3.14, true), stateMachine.Parameters.GetNextParameters<int, string, double, bool>());
         Assert.True(stateMachine.Parameters.IsPreviousSet);
         Assert.True(stateMachine.Parameters.IsNextSet);
-        Assert.Null(stateMachine.CurrentTransition);
     }
 
     [Fact]
