+++ tests/ZCrew.StateCraft.IntegrationTests/ExceptionHandling/TransitionExceptionTests{T1,T2,T3}.cs
@@ -36,7 +36,6 @@ public class TransitionExceptionTests_T1_T2_T3
         Assert.Empty(stateMachine.Parameters.CurrentParameterTypes);
         Assert.False(stateMachine.Parameters.IsPreviousSet);
         Assert.False(stateMachine.Parameters.IsNextSet);
-        Assert.Null(stateMachine.CurrentTransition);
     }
 
     [Fact]
@@ -142,7 +141,6 @@ public class TransitionExceptionTests_T1_T2_T3
         Assert.Empty(stateMachine.Parameters.CurrentParameterTypes);
         Assert.False(stateMachine.Parameters.IsPreviousSet);
         Assert.False(stateMachine.Parameters.IsNextSet);
-        Assert.Null(stateMachine.CurrentTransition);
     }
 
     [Fact]
@@ -252,7 +250,6 @@ public class TransitionExceptionTests_T1_T2_T3
         Assert.Empty(stateMachine.Parameters.CurrentParameterTypes);
         Assert.False(stateMachine.Parameters.IsPreviousSet);
         Assert.False(stateMachine.Parameters.IsNextSet);
-        Assert.Null(stateMachine.CurrentTransition);
     }
 
     [Fact]
@@ -376,7 +373,6 @@ public class TransitionExceptionTests_T1_T2_T3
         Assert.Equal((42, "hello", 3.14), stateMachine.Parameters.GetNextParameters<int, string, double>());
         Assert.True(stateMachine.Parameters.IsPreviousSet);
         Assert.True(stateMachine.Parameters.IsNextSet);
-        Assert.Null(stateMachine.CurrentTransition);
     }
 
     [Fact]
