+++ tests/ZCrew.StateCraft.IntegrationTests/ExceptionHandling/TransitionExceptionTests{T}.cs
@@ -36,7 +36,6 @@ public class TransitionExceptionTests_T
         Assert.Equal(42, stateMachine.Parameters.GetCurrentParameter<int>());
         Assert.False(stateMachine.Parameters.IsPreviousSet);
         Assert.False(stateMachine.Parameters.IsNextSet);
-        Assert.Null(stateMachine.CurrentTransition);
     }
 
     [Fact]
@@ -145,7 +144,6 @@ public class TransitionExceptionTests_T
         Assert.Equal(42, stateMachine.Parameters.GetCurrentParameter<int>());
         Assert.False(stateMachine.Parameters.IsPreviousSet);
         Assert.False(stateMachine.Parameters.IsNextSet);
-        Assert.Null(stateMachine.CurrentTransition);
     }
 
     [Fact]
@@ -251,7 +249,6 @@ public class TransitionExceptionTests_T
         Assert.Equal(42, stateMachine.Parameters.GetCurrentParameter<int>());
         Assert.False(stateMachine.Parameters.IsPreviousSet);
         Assert.False(stateMachine.Parameters.IsNextSet);
-        Assert.Null(stateMachine.CurrentTransition);
     }
 
     [Fact]
@@ -357,7 +354,6 @@ public class TransitionExceptionTests_T
         Assert.Equal("hello", stateMachine.Parameters.GetNextParameter<string>());
         Assert.True(stateMachine.Parameters.IsPreviousSet);
         Assert.True(stateMachine.Parameters.IsNextSet);
-        Assert.Null(stateMachine.CurrentTransition);
     }
 
     [Fact]
