+++ tests/ZCrew.StateCraft.IntegrationTests/ExceptionHandling/TryTransitionExceptionTests{T1,T2,T3}.cs
@@ -33,7 +33,6 @@ public class TryTransitionExceptionTests_T1_T2_T3
         Assert.Empty(stateMachine.Parameters.CurrentParameterTypes);
         Assert.False(stateMachine.Parameters.IsPreviousSet);
         Assert.False(stateMachine.Parameters.IsNextSet);
-        Assert.Null(stateMachine.CurrentTransition);
     }
 
     [Fact]
@@ -153,7 +152,6 @@ public class TryTransitionExceptionTests_T1_T2_T3
         Assert.Empty(stateMachine.Parameters.CurrentParameterTypes);
         Assert.False(stateMachine.Parameters.IsPreviousSet);
         Assert.False(stateMachine.Parameters.IsNextSet);
-        Assert.Null(stateMachine.CurrentTransition);
     }
 
     [Fact]
@@ -274,7 +272,6 @@ public class TryTransitionExceptionTests_T1_T2_T3
         Assert.Empty(stateMachine.Parameters.CurrentParameterTypes);
         Assert.False(stateMachine.Parameters.IsPreviousSet);
         Assert.False(stateMachine.Parameters.IsNextSet);
-        Assert.Null(stateMachine.CurrentTransition);
     }
 
     [Fact]
@@ -410,7 +407,6 @@ public class TryTransitionExceptionTests_T1_T2_T3
         Assert.Equal((42, "hello", true), stateMachine.Parameters.GetNextParameters<int, string, bool>());
         Assert.True(stateMachine.Parameters.IsPreviousSet);
         Assert.True(stateMachine.Parameters.IsNextSet);
-        Assert.Null(stateMachine.CurrentTransition);
     }
 
     [Fact]
@@ -486,7 +482,6 @@ public class TryTransitionExceptionTests_T1_T2_T3
         Assert.Empty(stateMachine.Parameters.CurrentParameterTypes);
         Assert.False(stateMachine.Parameters.IsPreviousSet);
         Assert.False(stateMachine.Parameters.IsNextSet);
-        Assert.Null(stateMachine.CurrentTransition);
     }
 
     [Fact]
