+++ tests/ZCrew.StateCraft.IntegrationTests/ExceptionHandling/TryTransitionExceptionTests{T1,T2}.cs
@@ -33,7 +33,6 @@ public class TryTransitionExceptionTests_T1_T2
         Assert.Empty(stateMachine.Parameters.CurrentParameterTypes);
         Assert.False(stateMachine.Parameters.IsPreviousSet);
         Assert.False(stateMachine.Parameters.IsNextSet);
-        Assert.Null(stateMachine.CurrentTransition);
     }
 
     [Fact]
@@ -141,7 +140,6 @@ public class TryTransitionExceptionTests_T1_T2
         Assert.Empty(stateMachine.Parameters.CurrentParameterTypes);
         Assert.False(stateMachine.Parameters.IsPreviousSet);
         Assert.False(stateMachine.Parameters.IsNextSet);
-        Assert.Null(stateMachine.CurrentTransition);
     }
 
     [Fact]
@@ -249,7 +247,6 @@ public class TryTransitionExceptionTests_T1_T2
         Assert.Empty(stateMachine.Parameters.CurrentParameterTypes);
         Assert.False(stateMachine.Parameters.IsPreviousSet);
         Assert.False(stateMachine.Parameters.IsNextSet);
-        Assert.Null(stateMachine.CurrentTransition);
     }
 
     [Fact]
@@ -358,7 +355,6 @@ public class TryTransitionExceptionTests_T1_T2
         Assert.Equal((42, "hello"), stateMachine.Parameters.GetNextParameters<int, string>());
         Assert.True(stateMachine.Parameters.IsPreviousSet);
         Assert.True(stateMachine.Parameters.IsNextSet);
-        Assert.Null(stateMachine.CurrentTransition);
     }
 
     [Fact]
@@ -431,7 +427,6 @@ public class TryTransitionExceptionTests_T1_T2
         Assert.Empty(stateMachine.Parameters.CurrentParameterTypes);
         Assert.False(stateMachine.Parameters.IsPreviousSet);
         Assert.False(stateMachine.Parameters.IsNextSet);
-        Assert.Null(stateMachine.CurrentTransition);
     }
 
     [Fact]
